requirements.txt
altgraph
annotated-types
anyio
autogen
certifi
cffi
charset-normalizer
cryptography
Cython
diskcache
distro
docker
FLAML
h11
httpcore
httpx
idna
jiter
joblib
numpy
openai
packaging
pandas
pyarmor
pyarmor.cli.core
pycparser
pydantic
pydantic_core
pyinstaller
pyinstaller-hooks-contrib
python-dateutil
python-dotenv
pytz
regex
requests
scikit-learn
scipy
six
sniffio
termcolor
threadpoolctl
tiktoken
tqdm
typing_extensions
tzdata
urllib3
google-generativeai
autogen-agentchat[gemini]~=0.2
jupyterlab